+++ generator/test/error_test_data/illigal_constructor.dart
@@ -1,0 +1,11 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'illigal_constructor.g.dart';
+
+@RealmModel()
+class _Bad {
+  @PrimaryKey()
+  late int id;
+
+  _Bad(this.id);
+}