+++ generator/test/error_test_data/invalid_model_name_prefix.dart
@@ -1,0 +1,8 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'invalid_model_name_prefix.g.dart';
+
+@RealmModel()
+class Bad {
+  // missing _ or $ prefix
+}