+++ generator/test/error_test_data/nullable_list_elements.dart
@@ -1,0 +1,15 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'nullable_list_elements.g.dart';
+
+@RealmModel()
+class _Other {}
+
+@RealmModel()
+class _Bad {
+  @PrimaryKey()
+  late int id;
+
+  late List<int?> okay;
+  late List<_Other?> wrong;
+}