+++ generator/test/good_test_data/list_initialization.dart
@@ -1,0 +1,8 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'list_initialization.g.dart';
+
+@RealmModel()
+class _Person {
+  late List<_Person> children;
+}