+++ generator/test/good_test_data/optional_argument.dart
@@ -1,0 +1,8 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'optional_argument.g.dart';
+
+@RealmModel()
+class _Person {
+  _Person? spouse;
+}