+++ generator/test/good_test_data/required_argument.dart
@@ -1,0 +1,9 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'required_argument.g.dart';
+
+@RealmModel()
+class _Person {
+  @PrimaryKey()
+  late String name;
+}