+++ CHANGELOG.md
@@ -5,6 +5,8 @@ x.x.x Release notes (yyyy-MM-dd)
 
 ### Enhancements
 * Support disableFormatUpgrade ([#310](https://github.com/realm/realm-dart/pull/310))
+* Support result value from write transaction callbacks ([#294](https://github.com/realm/realm-dart/pull/294/))
+
 
 0.2.1+alpha Release notes (2022-03-20)
 ==============================================================